# IDGS901_API_Balones/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BalonEvo",
    "version": "v1"
  },
  "paths": {
...
    "/api/Productos": {
      "get": {
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Productos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Productos"
            }
          }
        }
      }
    },
    "/api/Productos/{id}": {
      "put": {
        "tags": [
          "Productos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Productos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos/{nombre}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "nombre",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Productos/buscarbyid/{id}": {
      "get": {
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Proveedor": {
      "get": {
        "tags": [
          "Proveedor"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Proveedor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Proveedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Proveedor"
            }
          }
        }
      }
    },
    "/api/Proveedor/{id}": {
      "put": {
        "tags": [
          "Proveedor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Proveedor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Proveedor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Productos": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "stock": {
          "format": "int32",
          "type": "integer"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
    "Proveedor": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "empresa": {
          "type": "string"
        },
        "rfc": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
...
  }
}